Move flow walkthrough to constants
